refactor: redis pub/sub 설정 인터페이스 분리
