+++ qootalk-server/module-infrastructure/src/main/java/com/lrchan/qootalk/infrastructure/messaging/redis/MessagePublisher.java
@@ -1,0 +1,7 @@
+package com.lrchan.qootalk.infrastructure.messaging.redis;
+
+import org.springframework.data.redis.listener.ChannelTopic;
+
+public interface MessagePublisher {
+    void publish(ChannelTopic topic, Object payload);
+}
